apps/ocean-marketplace/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
bitarray==1.2.2
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coloredlogs==15.0.1
cryptography==3.3.2
cytoolz==0.11.2
datasets==1.17.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
enforce-typing==1.0.0.post1
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.4.2
frozenlist==1.2.0
fsspec==2021.11.1
hexbytes==0.2.2
huggingface-hub==0.2.1
humanfriendly==10.0
idna==3.3
ipfshttpclient==0.7.0a1
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json-sempai==0.4.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lru-dict==1.1.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==5.2.0
multiprocess==0.70.12.2
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
notebook==6.4.6
numpy==1.22.0
ocean-contracts==0.6.9
ocean-lib==0.8.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.12.0
Pygments==2.11.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.0
rlp==2.0.1
sacremoses==0.0.46
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tokenizers==0.10.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.7
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.2
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2